# api-docs.yaml
openapi: 3.0.1
info:
  title: Instant Messaging API
  description: |
    API REST pour une messagerie instantanée qui comprend la gestion  des utilisateur·rice·s, des conversations et des messages. Cette API est utilisée dans  le cadre des modules 194, 195, 322, et 426 à l'EPAI Fribourg.
  contact:
    name: Jérôme Frossard
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: http://localhost:8080/api/v1
  description: Generated server url
security:
- BearerAuth: []
...
paths:
...
  /conversations:
    get:
      tags:
      - conversation
      summary: Récupère les conversations de l'utilisateur·rice.
      description: Récupère les conversations dans lesquelles l'utilisateur·rice connecté·e
        est actif·ve.
      operationId: get-conversations
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "429":
          description: Too Many Requests
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "415":
          description: Unsupported Media Type
        "200":
          description: Liste de conversations récupérée avec succès.
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/ConversationsReponseDto"
              example:
                _embedded:
                  conversations:
                  - id: 100
                    group: false
                    _embedded:
                      participants:
                      - userId: 11
                        username: sheana@example.com
                        role: MEMBER
                        status: ACTIVE
                        _links:
                          self:
                            href: /api/v1/conversations/100/participants/11
                      - userId: 12
                        username: idaho@example.com
                        role: MEMBER
                        status: ACTIVE
                        _links:
                          self:
                            href: /api/v1/conversations/100/participants/12
                    _links:
                      self:
                        href: /api/v1/conversations/100
                      participants:
                        href: /api/v1/conversations/100/participants
                _links:
                  self:
                    href: /api/v1/conversations?page=0&size=20
                page:
                  size: 20
                  totalElements: 1
                  totalPages: 1
                  number: 0
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
    post:
      tags:
      - conversation
      summary: Crée une nouvelle conversation.
      description: "Crée une nouvelle conversation avec un ou plusieurs participant·\
        e·s. Le nom d'utilisateur peut être soit une adresse e-mail, soit un numé\
        ro de téléphone au format E.164."
      operationId: create-conversation
      requestBody:
        description: Utilisateur·rice à créer
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationCreateDto"
            examples:
              Nouvelle conversation privée (username):
                description: Nouvelle conversation privée (username)
                value:
                  username: siona@example.com
              Nouvelle conversation de groupe (tableau de participants):
                description: Nouvelle conversation de groupe (tableau de participants)
                value:
                  participants:
                  - username: siona@example.com
                  - userId: 15
        required: true
      responses:
        "429":
          description: Too Many Requests
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "415":
          description: Unsupported Media Type
        "406":
          description: Not Acceptable
        "201":
          description: Conversation a été créée avec succès.
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/ConversationResponseDto"
              examples:
                Conversation privée:
                  description: Conversation privée
                  value:
                    id: 1003
                    group: false
                    _embedded:
                      participants:
                      - userId: 13
                        username: siona@example.com
                        role: MEMBER
                        status: ACTIVE
                        _links:
                          self:
                            href: /api/v1/conversations/1003/participants/13
                      - userId: 11
                        username: sheana@example.com
                        role: MEMBER
                        status: ACTIVE
                        _links:
                          self:
                            href: /api/v1/conversations/1003/participants/11
                    _links:
                      self:
                        href: /api/v1/conversations/1003
                      participants:
                        href: /api/v1/conversations/1003/participants
                Conversation de groupe:
                  description: Conversation de groupe
                  value:
                    id: 1003
                    group: true
                    _embedded:
                      participants:
                      - userId: 13
                        username: siona@example.com
                        role: MEMBER
                        status: ACTIVE
                        _links:
                          self:
                            href: /api/v1/conversations/1003/participants/13
                      - userId: 15
                        username: fallom@example.com
                        role: MEMBER
                        status: ACTIVE
                        _links:
                          self:
                            href: /api/v1/conversations/1003/participants/15
                      - userId: 11
                        username: sheana@example.com
                        role: OWNER
                        status: ACTIVE
                        _links:
                          self:
                            href: /api/v1/conversations/1003/participants/11
                    _links:
                      self:
                        href: /api/v1/conversations/1003
                      participants:
                        href: /api/v1/conversations/1003/participants
        "500":
          description: Internal Server Error
...
  /conversations/{conversationId}:
    get:
      tags:
      - conversation
      summary: Récupère les détails d'une conversation.
      description: Récupère les détails d'une conversation à partir de son identifiant.
      operationId: get-conversation
      parameters:
      - name: conversationId
        in: path
        description: Identifiant de la conversation
        required: true
        schema:
          type: integer
          format: int64
        example: 100
      responses:
        "429":
          description: Too Many Requests
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "200":
          description: Détails de la conversation récupérée avec succès.
          content:
            application/hal+json:
              schema:
                $ref: "#/components/schemas/ConversationResponseDto"
              example:
                id: 100
                group: false
                _embedded:
                  participants:
                  - userId: 11
                    username: sheana@example.com
                    role: MEMBER
                    status: ACTIVE
                    _links:
                      self:
                        href: http://localhost:8080/api/v1/conversation/100/participants/11
                  - userId: 12
                    username: idaho@example.com
                    role: MEMBER
                    status: ACTIVE
                    _links:
                      self:
                        href: http://localhost:8080/api/v1/conversation/100/participants/12
                _links:
                  self:
                    href: http://localhost:8080/api/v1/conversations/100
                  participants:
                    href: http://localhost:8080/api/v1/conversation/100/participants
        "415":
          description: Unsupported Media Type
        "406":
          description: Not Acceptable
        "500":
          description: Internal Server Error
...
components:
  schemas:
...
    Links:
      type: object
      additionalProperties:
        $ref: "#/components/schemas/Link"
...
    ConversationCreateDto:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        participants:
          type: array
          items:
            $ref: "#/components/schemas/ParticipantDto"
    ParticipantDto:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        role:
          type: string
        status:
          type: string
    ConversationResponseDto:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            participants:
              type: array
              items:
                $ref: "#/components/schemas/ParticipantResponseDto"
        _links:
          $ref: "#/components/schemas/Links"
        id:
          type: integer
          format: int64
        group:
          type: boolean
    ParticipantResponseDto:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        role:
          type: string
        status:
          type: string
        _links:
          $ref: "#/components/schemas/Links"
...
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
...
    ConversationsReponseDto:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            conversations:
              type: array
              items:
                $ref: "#/components/schemas/ConversationResponseDto"
        _links:
          $ref: "#/components/schemas/Links"
        page:
          $ref: "#/components/schemas/PageMetadata"
    Link:
      type: object
      properties:
        href:
          type: string
        hreflang:
          type: string
        title:
          type: string
        type:
          type: string
        deprecation:
          type: string
        profile:
          type: string
        name:
          type: string
        templated:
          type: boolean
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer